requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
djangorestframework
django-safedelete==1.3.1
python-dotenv==0.21.0